upload policy

+++ app/Http/Controllers/Upload/UploadController.php
@@ -38,6 +38,7 @@ class UploadController extends Controller
     public function destroy(File $file, Upload $upload)
     {
       $this->authorize('touch', $file);
+      $this->authorize('touch', $upload);
 
       $upload->delete();
     }
